+++ src/page_entry.c
@@ -74,7 +74,7 @@ int compare_page_entry_by_page_id(const void* page_ent1, const void* page_ent2)
 	return compare_page_id(((page_entry*)page_ent1)->page_id, ((page_entry*)page_ent2)->page_id);
 }
 
-unsigned long long int hash_page_entry_by_page_id(const void* page_ent)
+unsigned int hash_page_entry_by_page_id(const void* page_ent)
 {
 	return hash_page_id(((page_entry*)page_ent)->page_id);
 }
@@ -85,7 +85,7 @@ int compare_page_entry_by_page_memory(const void* page_ent1, const void* page_en
 							(uintptr_t)(((page_entry*)page_ent2)->page_memory));
 }
 
-unsigned long long int hash_page_entry_by_page_memory(const void* page_ent)
+unsigned int hash_page_entry_by_page_memory(const void* page_ent)
 {
 	return ((uintptr_t)((page_entry*)page_ent)->page_memory) / 512;
 }
