+++ src/page_id_helper_functions.c
@@ -1,6 +1,6 @@
 #include<page_id_helper_functions.h>
 
-unsigned long long int hash_page_id(PAGE_ID page_id)
+unsigned int hash_page_id(PAGE_ID page_id)
 {
 	return ((page_id*2654435761)|(page_id*131));
 }
